+++ ui/v2.5/src/plugins.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { PatchFunction } from "./pluginApi";
+import { PatchFunction } from "./patch";
 
 export const PluginRoutes: React.FC<React.PropsWithChildren<{}>> =
   PatchFunction("PluginRoutes", (props: React.PropsWithChildren<{}>) => {
